// src/packages/types/report/Ser_ReportRoVarianceCost.ts
// export interface SerInventoryReportInOutBalance {
//   PartCode: string; //MaPhuTung
//   VieName: string; //TenHang
//   Unit: string; //DonViTinh
//   SLD: string; //SLD
//   TGD: string; //TGD
//   SLN: string; //SLN
//   TGN: string; //TGN
//   SLX: string; //SLX
//   TGX: string; //TGX
//   SLC: string; //SLC
//   TGC: string; //TGC
//   Location: string; //KhoViTri
//   lst_Ser_InventoryReport_InOutBalance: any;
// }

export interface Search_Ser_ReportRoVarianceCost_Params {
  FromDate: any
  ToDate: any
  DealerCode?: string;
  FlagDataWH: any
  FromDateFromTo: any;
  FlagStockOut: any
  FlagQueryFullHistory: any
  RoNo: any
}


// public object RoID { get; set; }
// public object RoNo { get; set; } // Số báo giá
// public object CheckInDate { get; set; } // Ngày vào xưởng
// public object ActualDeliveryDate { get; set; } // Ngày giao xe
// public object Cusid { get; set; }
// public object CarID { get; set; }
// public object PlateNo { get; set; } // Biển số xe
// public object TradeMarkCode { get; set; }
// public object ModelName { get; set; }  // Model
// public object IDPartSer { get; set; }
// public object PaSeCode { get; set; } // Mã PT-CV
// public object PaSeName { get; set; } // Tên PT-CV
// public object GiaBan { get; set; } // Giá Bán
// public object Quantity { get; set; } // Số lượng
// public object VATBan { get; set; } // VAT
// public object Factor { get; set; } // Hệ số
// public object GiaChuan { get; set; } // Giá chuẩn
// public object VATChuan { get; set; } // VAT chuẩn
// public object TriGia { get; set; } // Giá trị
// public object varianceCost { get; set; } // Giá trị chênh lệch";
// public object SPIndex { get; set; }
// public object Creator { get; set; }
// public object su_UserCode { get; set; } // CVDV
// public object su_UserName { get; set; } // Khách hàng
// public object TriGia01 { get; set; }
// public object ChenhLech { get; set; } // Chênh lệch
// public object SoLuong { get; set; } // Số lượng